+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Features
   - Add a constructor without `NativeAdUnit` in `CriteoNativeLoader` for InHouse integration
   - Add API to collect different levels of signals which will be used to bid based on context
+  - Add `Criteo.Builder#debugLogsEnabled` to enable useful logs to troubleshoot the SDK
 
 # Version 4.0.0